move entity -> model -> entity mapper to correct place
