+++ src/test/java/com/guillaume/training/controller/mapper/ExerciceMapperTest.java
@@ -2,11 +2,10 @@ package com.guillaume.training.controller.mapper;
 
 import com.guillaume.training.controller.dto.ExercicePayload;
 import com.guillaume.training.controller.dto.ExerciceResponse;
-import com.guillaume.training.repository.entity.ExerciceEntity;
 import com.guillaume.training.service.model.Exercice;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class ExerciceMapperTest {
     @Test
@@ -14,24 +13,6 @@ class ExerciceMapperTest {
         ExercicePayload exercice = new ExercicePayload(1L, "exercice Name", "description");
         Exercice expected = new Exercice(1L, "exercice Name", "description");
         Exercice actual = ExerciceMapper.getModelFromDTO(exercice);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void shouldMapCorrectlyAnExerciceModelToExerciceEntity(){
-        Exercice exercice = new Exercice(1L, "exercice Name", "description");
-        ExerciceEntity expected = new ExerciceEntity(1L, "exercice Name", "description");
-        ExerciceEntity actual = ExerciceMapper.getEntityFromModel(exercice);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void shouldMapCorrectlyAnExerciceEntityToExerciceModel(){
-        ExerciceEntity exercice = new ExerciceEntity(1L, "exercice Name", "description");
-        Exercice expected = new Exercice(1L, "exercice Name", "description");
-        Exercice actual = ExerciceMapper.getModelFromEntity(exercice);
 
         assertEquals(expected, actual);
     }
