+++ src/test/java/com/guillaume/training/controller/mapper/UserMapperTest.java
@@ -2,11 +2,10 @@ package com.guillaume.training.controller.mapper;
 
 import com.guillaume.training.controller.dto.UserPayload;
 import com.guillaume.training.controller.dto.UserResponse;
-import com.guillaume.training.repository.entity.UserEntity;
 import com.guillaume.training.service.model.User;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class UserMapperTest {
 
@@ -15,24 +14,6 @@ class UserMapperTest {
         UserPayload user = new UserPayload(1L, "User Name", "email@email.fr");
         User expected = new User(1L, "User Name", "email@email.fr");
         User actual = UserMapper.getModelFromDTO(user);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void shouldMapCorrectlyAnUserModelToUserEntity(){
-        User user = new User(1L, "User Name", "email@email.fr");
-        UserEntity expected = new UserEntity(1L, "User Name", "email@email.fr");
-        UserEntity actual = UserMapper.getEntityFromModel(user);
-
-        assertEquals(expected, actual);
-    }
-
-    @Test
-    public void shouldMapCorrectlyAnUserEntityToUserModel(){
-        UserEntity user = new UserEntity(1L, "User Name", "email@email.fr");
-        User expected = new User(1L, "User Name", "email@email.fr");
-        User actual = UserMapper.getModelFromEntity(user);
 
         assertEquals(expected, actual);
     }
